// frontend/src/lib/protocol/generated/PublishedTrafficTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LatLon } from "./LatLon";
import type { TrafficAlarmLevel } from "./TrafficAlarmLevel";
import type { TrafficType } from "./TrafficType";

export type PublishedTrafficTarget = { id: string, position: LatLon, altitudeMslMeters: number | null, trafficType: TrafficType, trackDegrees: number | null, alarmLevel: TrafficAlarmLevel, stale: boolean, };